refactor: rework triangle input
Replaced triangle input to be more streamlined with loops instead of
repetitive code.

Improved the maintainability while decreasing the risk of errors
that could be made in the future by future maintainers by lowering
the amount of code for creating a triangle. Other methods received
changes to fit with the metaclass creation.