docs: update README.md program instructions
Separated the uninstalling of Windows and Linux in the README.md to
make it simpler for users of either platform to use the program for
their own use.

The instructions were also updated with the reorganization of the file
paths to properly be used by the user. In addition, some white space
was removed along with periods at the version of languages required.